Create data models and upadte migrations

+++ db/migrations/20180104171635_create_hscodes.cr
@@ -10,13 +10,13 @@ class CreateHscodes::V20180104171635 < LuckyMigrator::Migration::V1
       add unit : String , default: "UN"
       add special_permission : String?
 
-      add duty : Int64, default: 0
-      add excise : Int64, default: 0
-      add vat : Int64, default: 15
-      add sur : Int64, default: 10
-      add withholding : Int64, default: 3
-      add ss_1 : Int64?
-      add ss_2 : Int64?
+      add duty : Int32, default: 0
+      add excise : Int32, default: 0
+      add vat : Int32, default: 15
+      add sur : Int32, default: 10
+      add withholding : Int32, default: 3
+      add ss_1 : Int32?
+      add ss_2 : Int32?
     end
 
     execute "CREATE INDEX hscodes_section_chapter_heading_index ON hscodes USING btree (section_id, chapter_id, heading_id);"
