+++ db/migrations/20180104171700_create_exports.cr
@@ -4,17 +4,17 @@ class CreateExports::V20180104171700 < LuckyMigrator::Migration::V1
       belongs_to Hscode, on_delete: :cascade
       belongs_to Destination, references: :countries, on_delete: :cascade
 
-      add year : Int64
-      add month : Int64?
+      add year : Int32
+      add month : Int32?
       add cpc : String?
-      add quantity : Int64?
+      add quantity : Int32?
 
-      add mass_gross_kg : Int64?
-      add mass_net_kg : Int64
-      add fob_etb_cents : Int64
-      add fob_usd_cents : Int64
-      add tax_etb_cents : Int64?
-      add tax_usd_cents : Int64?
+      add mass_gross_kg : Int32?
+      add mass_net_kg : Int32
+      add fob_etb_cents : Int32
+      add fob_usd_cents : Int32
+      add tax_etb_cents : Int32?
+      add tax_usd_cents : Int32?
     end
 
     execute "CREATE INDEX ON exports (year, month);"
